+++ tests/test_base.py
@@ -79,7 +79,7 @@ def test_displacement_property(movebank_relocations):
         index=pd.Index(["Habiba", "Salif Keita"], name="groupby_col"),
     )
     pd.testing.assert_series_equal(
-        trajectory.gdf.groupby("groupby_col")[trajectory.gdf.columns].apply(ecoscope.base.Trajectory.get_displacement),
+        trajectory.gdf.groupby("groupby_col")[trajectory.gdf.columns].apply(ecoscope.base.trajectory.get_displacement),
         expected,
     )
 
@@ -92,7 +92,7 @@ def test_tortuosity(movebank_relocations):
     )
     pd.testing.assert_series_equal(
         trajectory.gdf.groupby("groupby_col")[trajectory.gdf.columns].apply(
-            ecoscope.base.Trajectory.get_tortuosity, include_groups=False
+            ecoscope.base.trajectory.get_tortuosity, include_groups=False
         ),
         expected,
     )
